+++ packages/expo-modules-core/android/src/main/java/expo/modules/kotlin/views/ViewDefinitionBuilder.kt
@@ -7,6 +7,7 @@ import android.content.Context
 import android.util.Log
 import android.view.View
 import android.view.ViewGroup
+import expo.modules.core.ViewManager
 import expo.modules.kotlin.AppContext
 import expo.modules.kotlin.Promise
 import expo.modules.kotlin.exception.CodedException
@@ -334,7 +335,12 @@ class ViewDefinitionBuilder<T : View>(
         UnexpectedException(e)
       }
     )
-    return View(context)
+
+    return if (ViewGroup::class.java.isAssignableFrom(viewClass.java)) {
+      ErrorViewGroup(context)
+    } else {
+      View(context)
+    }
   }
 
   private fun getPrimaryConstructor(): KFunction<T>? {